APM-1744 Move createCacheKey to class and test
