+++ s3CacheTask/s3Utils/tests/s3CacheAction.test.js
@@ -1,7 +1,6 @@
 const AWS = require('aws-sdk');
 const path = require('path');
 const { S3CacheAction } = require('../s3CacheAction');
-const { createCacheKey } = require('../createCacheKey');
 
 const vars = {
     credentials: {
@@ -15,6 +14,31 @@ const vars = {
 };
 
 describe('S3CacheAction', () => {
+    describe('createCacheKey', () => {
+        const s3client = new S3CacheAction(new AWS.S3({
+            credentials: vars.credentials,
+            endpoint: vars.endpoint,
+            s3ForcePathStyle: true
+        }));
+
+        test('return string with same number of "/" separated parts', async () => {
+            const keyInput = '"foo" | foo/bar/foo | foo.txt';
+            const inputParts = keyInput.split('|').map(part => part.trim());
+            const keyOutput = await s3client.createCacheKey(keyInput, __dirname);
+            const outputParts = keyOutput.split('/').map(part => part.trim());
+    
+            expect(inputParts.length).toBe(outputParts.length);
+        });
+    
+        test('returns the same result on each call', async () => {
+            const keyInput = '"foo" | foo/bar/foo | foo.txt';
+            const firstCall = await s3client.createCacheKey(keyInput, __dirname);
+            const secondCall = await s3client.createCacheKey(keyInput, __dirname);
+    
+            expect(firstCall).toBe(secondCall);
+        });
+    });
+
     describe('createCacheEntry', () => {
         const s3client = new S3CacheAction(new AWS.S3({
             credentials: vars.credentials,
@@ -29,7 +53,7 @@ describe('S3CacheAction', () => {
         describe('happy path', () => {    
             test('successfully uploads file to s3 bucket.', async () => {
                 const targetPath = path.resolve(__dirname, 'testData/test.json');
-                const keyName = await createCacheKey('"test" | testData | testData/test.json', __dirname);
+                const keyName = await s3client.createCacheKey('"test" | testData | testData/test.json', __dirname);
     
                 const resp = await s3client.createCacheEntry(targetPath, vars.buckets.createBucket, keyName);
     
@@ -39,7 +63,7 @@ describe('S3CacheAction', () => {
     
             test('successfully uploads directory to s3 bucket.', async () => {
                 const targetPath = path.resolve(__dirname, 'testData');
-                const keyName = await createCacheKey(`"Test Data" | testData`, __dirname);
+                const keyName = await s3client.createCacheKey(`"Test Data" | testData`, __dirname);
     
                 const resp = await s3client.createCacheEntry(targetPath, vars.buckets.createBucket, keyName);
     
@@ -53,7 +77,7 @@ describe('S3CacheAction', () => {
                 test('missing targetPath parameter.', async () => {
                     try {
                         const targetPath = undefined;
-                        const keyName = await createCacheKey('"test" | testData | testData/test.json', __dirname);
+                        const keyName = await s3client.createCacheKey('"test" | testData | testData/test.json', __dirname);
     
                         await s3client.createCacheEntry(targetPath, vars.buckets.createBucket, keyName);
                     } catch (error) {
@@ -65,7 +89,7 @@ describe('S3CacheAction', () => {
                 test('invalid targetPath path.', async () => {
                     try {
                         const targetPath = 'not-a-real-path';
-                        const keyName = await createCacheKey('"test" | testData | testData/test.json', __dirname);
+                        const keyName = await s3client.createCacheKey('"test" | testData | testData/test.json', __dirname);
         
                         await s3client.createCacheEntry(targetPath, vars.buckets.createBucket, keyName);
                     } catch (error) {
@@ -79,7 +103,7 @@ describe('S3CacheAction', () => {
                     try {
                         const bucketName = 'bucket-doesnt-exist';
                         const targetPath = path.resolve(__dirname, 'testData/test.json');
-                        const keyName = await createCacheKey('"test" | testData | testData/test.json', __dirname);
+                        const keyName = await s3client.createCacheKey('"test" | testData | testData/test.json', __dirname);
         
                         await s3client.createCacheEntry(targetPath, bucketName, keyName);
                     } catch (error) {
@@ -91,7 +115,7 @@ describe('S3CacheAction', () => {
                     try {
                         const bucketName = undefined;
                         const targetPath = path.resolve(__dirname, 'testData/test.json');
-                        const keyName = await createCacheKey('"test" | testData | testData/test.json', __dirname);
+                        const keyName = await s3client.createCacheKey('"test" | testData | testData/test.json', __dirname);
         
                         await s3client.createCacheEntry(targetPath, bucketName, keyName);
                     } catch (error) {
